Fix for user_override setting firing unreliably.
Addresses #21

Currently, when the "+" option is set for a hook event. Only hooks were the user id is 1 will actually fire. This is because True translates to 1 when the ORM converts the query to SQL. This change should now have the correct user_override value set when find_and_fire_hook is called.
 
In an effort to move this issue forward I am pushing the change proposed in #21 without a unit test.